+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/PlatformApi.java
@@ -53,6 +53,7 @@ import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntityType;
 import org.apiguardian.api.API;
 
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.1.0")
@@ -164,6 +165,12 @@ public interface PlatformApi {
     EnergyItemHelper getEnergyItemHelper();
 
     EnergyStorage asItemEnergyStorage(EnergyStorage energyStorage, ItemStack stack);
+
+    EnergyStorage asBlockItemEnergyStorage(
+        EnergyStorage energyStorage,
+        ItemStack stack,
+        BlockEntityType<?> blockEntityType
+    );
 
     NetworkBoundItemHelper getNetworkBoundItemHelper();